Move pid locking manager and exception to fcrepo-clients package
