+++ fcrepo-clients/src/main/java/edu/unc/lib/dl/persist/api/services/PidLockManager.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package edu.unc.lib.dl.util;
+package edu.unc.lib.dl.persist.api.services;
 
 import com.google.common.util.concurrent.Striped;
 import edu.unc.lib.dl.fedora.PID;
